Fix devcontainer configuration to make IntelliSense work.
In case it still doesn't work, select the ".venv" folder in the Project Explorer of VS Code.
From the contextual menu choose "Copy Path".
Click on the Python version on the bottom-right of the status bar.
Choose "Enter interpreter path" and paste the copied path (only the path to the ".venv" folder, not the complete path to the python bin).
From now on IntelliSense should work (imports are resolved inside the VS Code Python editor windows).